refactor: move testable class to mock
